parse_rdata: Treat rdf_uname the same as rdf_name
This commit treats rdf_name's and rdf_name's exactly the same when
decompressing domain names embedded in rdata. This affects the parsing
of DNAME, A6, KX, RRSIG, and NSEC RR's, which are currently defined in
record_descr with rdf_uname's. The use of rdf_uname in the record_descr
definitions indicates that those RRs are specified as being generated
without compression pointers.

However, the DNS specifications only require that *generators* of these
RR's must not use compression pointers. Consumers are not required to
enforce this restriction, which allows bad generator code to persist.

This change should result in more lenient parsing of RRSIG and NSEC
records in particular, as some commonly deployed proprietary DNSSEC
implementations incorrectly generate these records with compression
pointers.